requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
sqlmodel==0.0.14
sqlalchemy==2.0.23
pydantic==2.5.0
pandas==2.1.3
python-constraint==1.4.0
requests==2.31.0
python-dotenv==1.0.0
